+++ src/main/java/block/meta/BlockInfo.java
@@ -1,7 +1,7 @@
 package block.meta;
 
 public record BlockInfo(
-        int id, // TODO check if we actually need this
+        byte id, // TODO check if we actually need this
         String name,
         // store texture indices for each side of a given block id
         // +X, -X, +Y, -Y, +Z, -Z
